Recursive deleting can be dangerous
On deleting files (videos) or SHM-Entries there is no need to use the -r option of "rm". It could be very dangerous, because in many setups Shinobi runs as root and could destroy in worst cast the whole file system of the host (e.g. in case of software bugs, where s.file("delete", "/") is called.